mcp/pyproject.toml
[project]
name = "kaiken-mcp"
version = "0.1.0"
description = "MCP Server para Query Builder - Kaiken Lab"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.23.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
